improved node size behavior

+++ empower-engine/src/value.rs
@@ -2,6 +2,6 @@
 #[derive(PartialEq, PartialOrd, Clone)]
 pub enum Value
 {
-    Integer(i32),
-    Float(f32),
+    Integer,
+    Float,
 }
